Create `Storer.go`, `MapStore.go` &  `SyncStore.go`
